update headers - all table
