+++ rubix-ui/frontend/src/components/pc/scanner/table.tsx
@@ -1,38 +1,24 @@
 import { useEffect, useState } from "react";
 import { Spin } from "antd";
-import { CreateModal } from "./create";
 import { openNotificationWithIcon } from "../../../utils/utils";
 import { Scanner } from "../../../../wailsjs/go/main/App";
+import { SCANNER_HEADERS } from "../../../constants/headers";
 import RbTable from "../../../common/rb-table";
 import { RbAddButton, RbRefreshButton } from "../../../common/rb-table-actions";
+import { CreateModal } from "./create";
 
 const ScannerTable = (props: any) => {
   let { data, isFetching, setSelectedIpPorts } = props;
 
   if (!data) return <></>;
 
+  const columns = SCANNER_HEADERS;
+
   const rowSelection = {
     onChange: (selectedRowKeys: any, selectedRows: any) => {
       setSelectedIpPorts(selectedRows);
     },
   };
-
-  const columns = [
-    {
-      title: "Ip",
-      dataIndex: "ip",
-      key: "ip",
-    },
-    {
-      title: "Port",
-      dataIndex: "ports",
-      render: (services: any[]) =>
-        services.map((service, index) => (
-          <p key={index}> {`${service.service}: ${service.port}`} </p>
-        )),
-      key: "ports",
-    },
-  ];
 
   return (
     <RbTable
